+++ src/elements/db/SegmentElementQuery.php
@@ -6,7 +6,6 @@
 namespace putyourlightson\campaign\elements\db;
 
 use craft\elements\db\ElementQuery;
-use craft\helpers\Db;
 use putyourlightson\campaign\elements\SegmentElement;
 use yii\db\Connection;
 
@@ -18,21 +17,6 @@ use yii\db\Connection;
 class SegmentElementQuery extends ElementQuery
 {
     /**
-     * @var string|null The segment type that the resulting segments must have.
-     */
-    public ?string $segmentType = null;
-
-    /**
-     * Sets the [[segmentType]] property.
-     */
-    public function segmentType(string $value): static
-    {
-        $this->segmentType = $value;
-
-        return $this;
-    }
-
-    /**
      * @inheritdoc
      */
     protected function beforePrepare(): bool
@@ -40,15 +24,8 @@ class SegmentElementQuery extends ElementQuery
         $this->joinElementTable('campaign_segments');
 
         $this->query->select([
-            'campaign_segments.segmentType',
             'campaign_segments.contactCondition',
-            'campaign_segments.conditions',
-            'campaign_segments.template',
         ]);
-
-        if ($this->segmentType) {
-            $this->subQuery->andWhere(Db::parseParam('campaign_segments.segmentType', $this->segmentType));
-        }
 
         return parent::beforePrepare();
     }