+++ src/records/SegmentRecord.php
@@ -9,10 +9,7 @@ use craft\db\ActiveRecord;
 
 /**
  * @property int $id
- * @property string $segmentType
  * @property array|null $contactCondition
- * @property array|null $conditions
- * @property string|null $template
  */
 class SegmentRecord extends ActiveRecord
 {